pyproject.toml
[tool.poetry]
name = "rates"
version = "0.1.0"
description = ""
authors = ["Timur Atambaev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "3.14"
psycopg2-binary = "2.9.8"
djangorestframework-simplejwt = "5.3.0"
drf-yasg = "1.20.0"
django-filter = "23.3"
django-cors-headers = "4.2.0"
django-redis = "5.3.0"
django-hosts = "5.2"
django-storages = "1.14.1"
django-redis-sessions = "0.6.2"
django-silk = "5.0.4"
django-cacheops = "7.0.1"
celery = "5.3.4"
redis = "5.0.1"
django-celery-beat = "2.5.0"
drf-spectacular = "0.26.5"
drf-spectacular-sidecar = "2023.9.1"
twilio = "8.9.0"
loguru = "0.7.2"
django-rest-passwordreset = "1.3.0"
python-decouple = "3.8"

[tool.poetry.dev-dependencies]
pytest = "7.4.2"
mypy = "^0.942"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = ">=3.8"
coverage = "^6.3.2"
flake8-logging-format = "^0.7.5"
flake8-alfred = "^1.1.1"
flake8-bugbear = "^22.3.23"
flake8-builtins = "^1.5.3"
flake8-coding = "^1.3.2"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.8.0"
flake8-debugger = "^4.0.0"
flake8-eradicate = "^1.2.0"
flake8-multiline-containers = "^0.0.18"
flake8-pep3101 = "^1.3.0"
flake8-print = "^4.0.0"
flake8-pytest-style = "^1.6.0"
flake8-return = "^1.1.3"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-SQL = "0.3.0"
pep8-naming = ">0.1"
flake8-cognitive-complexity = "^0.1.0"
flake8-isort = "^4.1.1"
flake8-annotations = "^2.8.0"
types-requests = "^2.27.16"
flake8-docstrings = "^1.6.0"
pytest-asyncio = "0.20.1"
pytest-env = "0.8.1"
factory-boy = "3.3.0"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.0.2"

[tool.black]
line-length = 79
include = '\.py$'

[tool.isort]
line_length = 79
skip="migrations"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
profile="black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
